Fixed switching to custom presets
It turns out, to switch to custom preset, you need to call the API with
{name: ..., bank: "My Presets"}. The value of bank is exposed in the
JSON RPC api when the list of available presets are fetched. This
changes the code so it works.

See: https://forum.modartt.com/viewtopic.php?pid=979806#p979806

Also, I've noted that the current_preset returned by the getInfo API
returns "My Presets/<PRESET NAME>" instead of just "<PRESET NAME>" for
custom presets. The way the code works now means that the custom preset
will be duplicated in the drop down menu (once as "<PRESET NAME>" and
once as "My Presets/<PRESET NAME>"). This is a minor problem so not
fixed here.